Time: 2 ms (50.69%) | Memory: 8.6 MB (65.28%) - LeetSync

+++ 9-palindrome-number/palindrome-number.cpp
@@ -3,7 +3,7 @@ public:
     bool isPalindrome(int x) {
         if(x<0) return false;
         int copy=x;
-        long long int rev=0;
+        long int rev=0;
         while(x>0){
             rev=rev*10+x%10;
             x/=10;
